Move checking if an addon is installed to its own file
